+++ src/data/store/arena/typed/impls/mod.rs
@@ -1,0 +1,4 @@
+// devela/src/data/store/arena/typed/impls/mod.rs
+
+mod array; // __arena_impl_array!
+mod vec; // __arena_impl_vec!