+++ src/data/store/arena/typed/mod.rs
@@ -1,14 +1,20 @@
 // devela/src/data/store/arena/typed/mod.rs
 //
-//!
+//! Typed monotonic arenas.
 //
 
-// mod define; // arena_typed!
+#[cfg(test)]
+mod _test;
+#[cfg(any(test, feature = "_docs_examples"))]
+mod _example;
+
+mod define; // arena!
+mod impls; // hidden macros for arena variants
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::{
-        //     define::*,
-        // };
+        pub use super::define::arena;
+        #[cfg(any(test, feature = "_docs_examples"))]
+        pub use super::_example::*;
     }
 }